+++ crates/polars-core/src/chunked_array/iterator/mod.rs
@@ -421,48 +421,6 @@ impl<T: PolarsObject> ObjectChunked<T> {
     }
 }
 
-// Make sure to call `rechunk` first!
-#[cfg(feature = "dtype-struct")]
-impl<'a> IntoIterator for &'a StructChunked {
-    type Item = &'a [AnyValue<'a>];
-    type IntoIter = StructIter<'a>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        let field_iter = self.fields().iter().map(|s| s.iter()).collect();
-
-        StructIter {
-            field_iter,
-            buf: vec![],
-        }
-    }
-}
-
-#[cfg(feature = "dtype-struct")]
-pub struct StructIter<'a> {
-    field_iter: Vec<SeriesIter<'a>>,
-    buf: Vec<AnyValue<'a>>,
-}
-
-#[cfg(feature = "dtype-struct")]
-impl<'a> Iterator for StructIter<'a> {
-    type Item = &'a [AnyValue<'a>];
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.buf.clear();
-
-        for it in &mut self.field_iter {
-            self.buf.push(it.next()?);
-        }
-        // SAFETY:
-        // Lifetime is bound to struct, we just cannot set the lifetime for the iterator trait
-        unsafe {
-            Some(std::mem::transmute::<&'_ [AnyValue], &'a [AnyValue]>(
-                &self.buf,
-            ))
-        }
-    }
-}
-
 #[cfg(feature = "dtype-struct")]
 impl<'a> IntoIterator for &'a StructChunked2 {
     type Item = Option< &'a [AnyValue<'a>]>;
