--- crates/polars-core/src/chunked_array/logical/struct_/from.rs
@@ -1,7 +1,0 @@
-use crate::prelude::*;
-
-impl DataFrame {
-    pub fn into_struct(self, name: &str) -> StructChunked {
-        StructChunked::new(name, &self.columns).unwrap()
-    }
-}
